requirements.txt
onnx==1.16.0
opencv-python==4.9.0.80
pandas==2.0.3
Pillow==9.5.0
pycocotools==2.0.7
PyQt5==5.15.9
pyqt5-tools==5.15.9.3.3
PyYAML
scikit-image==0.21.0
scikit-learn==1.3.2
torch
torchvision
tqdm==4.66.1